drm: panel: s6e63m0: Workaround screen corruption on boot
The panel driver gets probed before PMIC, so the regulator_bulk_get() call
fails with EPROBE_DEFER, making the driver to probe later in the boot
process. However, the screen gets corrupted in some time with noise
displayed. To resolve this, reset GPIO is now requested before the
regulator, setting the panel to the reset state.

Signed-off-by: Sergey Larin <cerg2010cerg2010@mail.ru>
Signed-off-by: Svyatoslav Ryhel <clamor95@gmail.com>