media: i2c: mt9m114: correct pad order for external devices
The MT9M114 driver exposes the IFP device's sink pad first, which causes
issues for external devices that rely on the OF graph to get the correct
pad information. This patch corrects the pad order by swapping the IFP's
source and sink pads, ensuring that the source pad is exposed first. This
allows external devices to correctly link with the IFP using OF graph.

Signed-off-by: Svyatoslav Ryhel <clamor95@gmail.com>